Change PMW3360_DATA struct data types to work with Teensy also

+++ src/PMW3360.cpp
@@ -446,9 +446,9 @@ PMW3360_DATA PMW3360::readBurst()
   uint8_t sl = burstBuffer[10];   // shutter LSB
   uint8_t sh = burstBuffer[11];   // shutter MSB
   
-  int x = xh<<8 | xl;
-  int y = yh<<8 | yl;
-  unsigned int shutter = sh<<8 | sl;
+  int16_t x = xh<<8 | xl;
+  int16_t y = yh<<8 | yl;
+  uint16_t shutter = sh<<8 | sl;
 
   data.isMotion = motion;
   data.isOnSurface = surface;
